requirements.txt
numpy==2.2.5
pip==25.1.1
pygame==2.6.1
PyOpenGL==3.1.9